Unify all floats to Decimal, integers to usize

+++ src/rest/mod.rs
@@ -395,9 +395,9 @@ impl Rest {
         &self,
         market: &str,
         side: OrderSide,
-        price: Option<f64>,
+        price: Option<Decimal>,
         r#type: OrderType,
-        size: f64,
+        size: Decimal,
         reduce_only: Option<bool>,
         ioc: Option<bool>,
         post_only: Option<bool>,
@@ -431,8 +431,8 @@ impl Rest {
     pub async fn modify_order(
         &self,
         order_id: usize,
-        price: Option<f64>,
-        size: Option<f64>,
+        price: Option<Decimal>,
+        size: Option<Decimal>,
         client_id: Option<&str>,
     ) -> Result<OrderInfo> {
         self.post(
